Add map metadata. Fix entity rendering.
Firstly, support for generic map metadata has been added. In the future,
this will allow any sort of non-tile data to be stored with maps. Right now,
it is being used for player spawn points.

Entities were previously being rendered incorrectly. The presence of a "magic"
offset number was evidence of this. The issue causing this has been resolved.