Add dynamic entity loading,  interacting with entities, chat bubbles.
This commit adds support for information about entities to be loaded from the map 
metadata so that they can be stored with a map. Details for doing so is in 
mapFormat.txt and addingEntities.txt. Interface method for a player spacebar
interacting with an entity is added too. Finally, a utility class for displaying speech
bubbles (and eventually chatboxes for cutscenes) is added preliminarily.